ui: some refactoring
